refactor: change component's naming
