+++ src/components/Avatar/index.css
@@ -2,7 +2,6 @@
   margin: 60px auto 40px;
   width: 320px;
   height: 400px;
-  
 }
 
 .avatar-container img {
