+++ src/components/Avatar/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import './Avatar.css';
+import './index.css';
 
-function Avatar(props) {  
+function Avatar(props) {
   return (
     <div className="avatar-container">
       <img src={props.data.avatar} alt={props.data.name} />
     </div>
   )
-}  
+}
 
 export default Avatar;
