+++ src/components/Description/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import './Description.css';
+import './index.css';
 
 function Description(props) {
   return (
-    <div className="description">    
+    <div className="description">
       <h1 className="name">{props.data.name}</h1>
       <p className="occupation">{props.data.occupation}</p>
       <p className="years">Годы жизни: <span>{props.data.birthDate}</span> - <span>{props.data.deathDate}</span></p>
