+++ src/components/Map/index.js
@@ -1,19 +1,19 @@
 import React, {Component} from 'react';
 import {Map, Marker, GoogleApiWrapper} from 'google-maps-react';
-import './Map.css'
+import './index.css'
 
-export class MapContainer extends Component {  
+export class MapContainer extends Component {
   render() {
     const {lat, lng} = this.props.data.locationsCoords;
     return (
-      <Map google={this.props.google}      
+      <Map google={this.props.google}
       initialCenter={{
             lat,
             lng
           }}
       className={'map'}
       zoom={12}>
-        <Marker          
+        <Marker
           position={{lat, lng}} />
       </Map>
     );
