+++ src/components/SearchPanel/index.css
